refactoring: host-app Nav component
NavBar 폴더로 이동
NavBarTab, NavBarTabs 로 파일 분리

propTypes 추가
